Split policy documents into three files.
Embedded documents that sit inside other policies, documents used to
issue another policy documents, or regular policy documents inside
the map server and DB.
